# packages/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "PAT or session token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Identity and access management API. Schema-driven, extensible, open source.",
    "title": "Zitadel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/login/flows/{id}": {
      "get": {
        "operationId": "getFlow",
        "parameters": [
          {
            "description": "Resource ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get login flow state"
          }
        },
        "summary": "Get login flow state",
        "tags": [
          "Auth"
        ]
      },
      "post": {
        "operationId": "submitFlow",
        "parameters": [
          {
            "description": "Resource ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Submit a login flow step"
          }
        },
        "summary": "Submit a login flow step",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "This instance",
      "url": "/"
    }
  ],
...
}